+++ resources/views/consultas_desarrollos/ver_desarrollos.blade.php
@@ -41,8 +41,12 @@
                             <td>{{$desarrollo-> sector}}</td>
                             <td>{{$desarrollo-> pais}}</td>
                             <td>
+                            @if(Session::get('tipo')== 'A')
                                <a  href="editar_desarrollo/{{$desarrollo->id}}"  type="button" class="btn btn-default btn-xs" ><span class="fa fa-pencil" data-toggle="tooltip" title="Da click para editar"></span></a>
-                                  <a  href="eliminar_desarrollo/{{$desarrollo->id}}"  type="button" class="btn btn-danger btn-xs" ><span class="fa fa-trash" data-toggle="tooltip" title="Da click para borrar"></span></a>
+                                <a  href="eliminar_desarrollo/{{$desarrollo->id}}"  type="button" class="btn btn-danger btn-xs" ><span class="fa fa-trash" data-toggle="tooltip" title="Da click para borrar"></span></a>
+                            @elseif(Session::get('tipo')== 'C')
+                              <a  href="editar_desarrollo/{{$desarrollo->id}}"  type="button" class="btn btn-default btn-xs" ><span class="fa fa-pencil" data-toggle="tooltip" title="Da click para editar"></span></a>
+                            @endif
                             </td>          
                         </tr>
                     </tbody>
