+++ resources/views/consultas_distincions/ver_distincions.blade.php
@@ -38,8 +38,12 @@
                             <td>{{$distincion-> pais}}</td>
                             <td>{{$distincion-> descripcion}}</td>
                             <td>
+                            @if(Session::get('tipo')=='A')
                                <a  href="editar_distincion/{{$distincion->id}}"  type="button" class="btn btn-default btn-xs" ><span class="fa fa-pencil" data-toggle="tooltip" title="Da click para editar"></span></a>
                                <a  href="eliminar_distincion/{{$distincion->id}}"  type="button" class="btn btn-danger btn-xs" ><span class="fa fa-trash" data-toggle="tooltip" title="Da click para borrar"></span></a>
+                            @elseif(Session::get('tipo')=='C')
+                               <a  href="editar_distincion/{{$distincion->id}}"  type="button" class="btn btn-default btn-xs" ><span class="fa fa-pencil" data-toggle="tooltip" title="Da click para editar"></span></a>
+                            @endif
                             </td>          
                         </tr>
                     </tbody>
