+++ resources/views/consultas_libros/ver_libros.blade.php
@@ -44,8 +44,12 @@
                             <td>{{$libro-> lista}}</td>
                             <td>{{$libro-> isbn}}</td>
                             <td>
+                            @if(Session::get('tipo') =='A')
                                <a  href="editar_libro/{{$libro->id}}"  type="button" class="btn btn-default btn-xs" ><span class="fa fa-pencil" data-toggle="tooltip" title="Da click para editar"></span></a>
                                <a  href="eliminar_libro/{{$libro->id}}"  type="button" class="btn btn-danger btn-xs" ><span class="fa fa-trash" data-toggle="tooltip" title="Da click para borrar"></span></a>
+                            @elseif(Session::get('tipo')=='C')
+                                <a  href="editar_libro/{{$libro->id}}"  type="button" class="btn btn-default btn-xs" ><span class="fa fa-pencil" data-toggle="tooltip" title="Da click para editar"></span></a>
+                            @endif
                             </td>          
                         </tr>
                     </tbody>
