+++ resources/views/consultas_resenias/ver_resenias.blade.php
@@ -46,8 +46,12 @@
                             <td>{{$resenia-> pos_autor}}</td>
                             <td>{{$resenia-> lista_autor}}</td>
                             <td>
+                            @if(Session::get('tipo')=='A')
                                <a  href="editar_resenias/{{$resenia->id}}"  type="button" class="btn btn-default btn-xs" ><span class="fa fa-pencil" data-toggle="tooltip" title="Da click para editar"></span></a>
                                <a  href="eliminar_resenia/{{$resenia->id}}"  type="button" class="btn btn-danger btn-xs" ><span class="fa fa-trash" data-toggle="tooltip" title="Da click para borrar"></span></a>
+                            @elseif(Session::get('tipo')=='C')
+                                <a  href="editar_resenias/{{$resenia->id}}"  type="button" class="btn btn-default btn-xs" ><span class="fa fa-pencil" data-toggle="tooltip" title="Da click para editar"></span></a>
+                            @endif
                             </td>          
                         </tr>
                     </tbody>
