+++ resources/views/consultas_trabajos/ver_trabajos.blade.php
@@ -39,8 +39,12 @@
                             <td>{{$trabajo-> lugar_evento}}</td>
                             <td>{{$trabajo-> entidad_evento}}</td>
                             <td>
+                            @if(Session::get('tipo')=='A')
                               <a  href="editar_trabajo/{{$trabajo->id}}" type="button" class="btn btn-default btn-xs" ><span class="fa fa-pencil" data-toggle="tooltip" title="Da click para editar"></span></a>
                               <a  href="eliminar_trabajo/{{$trabajo->id}}"  type="button" class="btn btn-danger btn-xs" ><span class="fa fa-trash" data-toggle="tooltip" title="Da click para borrar"></span></a>
+                            @elseif(Session::get('tipo')=='C')
+                            <a  href="editar_trabajo/{{$trabajo->id}}" type="button" class="btn btn-default btn-xs" ><span class="fa fa-pencil" data-toggle="tooltip" title="Da click para editar"></span></a>
+                            @endif
                             </td>          
                         </tr>
                     </tbody>
